+++ l10n.yaml
@@ -1,0 +1,3 @@
+arb-dir: lib/l10n/arb
+template-arb-file: app_en.arb
+output-localization-file: app_localizations.dart
