+++ lib/l10n/l10n.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/widgets.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+
+extension LocalizedContext on BuildContext {
+  AppLocalizations get l10n => AppLocalizations.of(this)!;
+}
