+++ pubspec.yaml
@@ -28,12 +28,19 @@ dependencies:
   flutter:
     sdk: flutter
 
+  # Dependency injection
+  get_it: ^6.1.1
+
+  # Internationalization
+  flutter_localizations:
+    sdk: flutter
+  intl: ^0.17.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  get_it: ^6.1.1
-
+  # Linting
   lint: ^1.5.3
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -44,6 +51,9 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+
+  # Flutter generation
+  generate: true
 
   # To add assets to your application, add an assets section, like this:
   assets:
